own2mqtt/requirements.txt
paho-mqtt==1.5.0
regex==2019.11.1
